feat(referral): add user ref circle withdrawal feature
Introduced a new endpoint to handle withdrawals for user ref circles.
Added necessary DTOs, a service method to process the withdrawal, and linked it with a queue for further processing. Enforced validations for minimum withdrawal amount and duplicate withdrawal prevention.

BREAKING CHANGE: Requires the `withdraw` queue and `MINIMUM_REF_WITHDRAW_AMOUNT` setting to be configured.